fixup do not try to delete unknown reports from database

+++ src/app/threat-dashboard/services/threat-report-overview.service.ts
@@ -124,6 +124,10 @@ export class ThreatReportOverviewService {
    *  @param id
    */
   public deleteThreatReport(id: string): Observable<ThreatReport[]> {
+    if (!id || id.trim().length === 0) {
+      return Observable.of([]);
+    }
+
     const url = this.reportsUrl + '/' + id;
     return Observable.create((observer) => {
       const sub  = this.genericService.delete(url).subscribe(
